Move more .d.ts files to .ts
