+++ packages/desktop-electron/index.ts
@@ -596,7 +596,7 @@ ipcMain.handle('relaunch', () => {
 });
 
 export type OpenFileDialogPayload = {
-  properties: OpenDialogSyncOptions['properties'];
+  properties?: OpenDialogSyncOptions['properties'];
   filters?: OpenDialogSyncOptions['filters'];
 };
