+++ packages/loot-core/src/platform/client/fetch/index.browser.ts
@@ -5,7 +5,7 @@ import { v4 as uuidv4 } from 'uuid';
 import { captureException, captureBreadcrumb } from '../../exceptions';
 import * as undo from '../undo';
 
-import type * as T from './index.d';
+import type * as T from './index-types';
 
 const replyHandlers = new Map();
 const listeners = new Map();
