+++ packages/loot-core/typings/vite-plugin-peggy-loader.ts
